requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.4
PyYAML==6.0.1
lxml==4.9.3
urllib3==2.0.7
fake-useragent==1.4.0